Export the Executor class for customization (#8)
* Export internal Executor class for experimentation

This class, similar to the Parser class, is exported only to assist people in implementing their own executors without duplicating too much code and should be used only as last resort for cases such as experimental syntax or if certain features could not be contributed upstream.

It is still part of the internal API (so far), so any changes to it are never considered breaking changes. The `graphql-executor` package must therefore be pinned to a specific patch version, e.g. `0.0.4`.

* Fix README