standardize iterator and async iterator methods
the two complete methods differ only in that the latter is async and has to handle promise rejection.
the two add methods differ substantially, in that the async method must
add the async iterator to the set of known iterators and call itself
recursively to add to the subsequentPayloads racing set.