Do not validate schema on first execution
Schemas can still be validated when desired using the dedicated validateSchema, which should usually be done as a separate step to avoid unnecessary revalidation.

Previously, by default, on first execution, a schema would be validated (if not validated already). This behavior could be disabled by using the assumeValid option with schema creation.