allow type resolution with GraphQLObjectTypes (#117)
...as opposed to just type names.

This reverts graphql#2905

This change was made upstream to support schema transformation functions like lexicographicallySortSchema which do not modify resolveType methods. Previously, a workaround was included that simply looks up the actual type from the type name returned by resolveType. In actuality, in my opinion, the schema transformation functions should natively or manually modify the resolveType function so it works correctly.

In any case, we restore the workaround here to preserve v14 support.